Move `getExportDatesString` to @utils
