+++ packages/web-config-server/src/export/ExportSurveyDataHandler.js
@@ -1,6 +1,6 @@
 import xlsx from 'xlsx';
 import fs from 'fs';
-import moment from 'moment';
+import { getExportDatesString } from '@tupaia/utils';
 
 import { requestFromTupaiaConfigServer } from './requestFromTupaiaConfigServer';
 import { USER_SESSION_CONFIG } from '/authSession';
@@ -57,8 +57,8 @@ export class ExportSurveyDataHandler extends RouteHandler {
 
     Object.entries(data.data).forEach(([surveyName, surveyData]) => {
       const header = surveyData.data.columns.length
-        ? `Data ${this.getExportDatesString(startDate, endDate)}`
-        : `No data for ${surveyName} ${this.getExportDatesString(startDate, endDate)}`;
+        ? `Data ${getExportDatesString(startDate, endDate)}`
+        : `No data for ${surveyName} ${getExportDatesString(startDate, endDate)}`;
 
       const titleAndHeaderData = [[`${reportTitle}: ${surveyName}`], [header]];
       // Using array of arrays (aoa) input as transformations like mergeSurveys
@@ -95,21 +95,4 @@ export class ExportSurveyDataHandler extends RouteHandler {
     xlsx.writeFile(workbook, filePath);
     this.res.download(filePath);
   }
-
-  getExportDatesString = (startDate, endDate) => {
-    const format = 'D-M-YY';
-    let dateString = '';
-
-    if (startDate && endDate) {
-      dateString = `between ${moment(startDate).format(format)} and ${moment(endDate).format(
-        format,
-      )} `;
-    } else if (startDate) {
-      dateString = `after ${moment(startDate).format(format)} `;
-    } else if (endDate) {
-      dateString = `before ${moment(endDate).format(format)} `;
-    }
-
-    return `${dateString}as of ${moment().format(format)}`;
-  };
 }
